Format / move user fixtures to conftest and save token session correctly
